requirements.txt
flask
flask-login
google-cloud-firestore
joblib
flask-mail
numpy
scikit-learn
